requirements.txt
Django
django-allauth
djangorestframework
django-rq
rq-scheduler
django-anymail
msal
argon2-cffi
django-axes
redis

